Web.Routing.Combinators

Documentation

data PathState

Constructors

Open

Closed

data Path (as :: [Type]) (pathState :: PathState) where

Constructors

Empty :: Path ('[] :: [Type]) 'Open

StaticCons :: forall (as :: [Type]) (pathState :: PathState). Text -> Path as pathState -> Path as pathState

VarCons :: forall a (as1 :: [Type]) (pathState :: PathState). (FromHttpApiData a, Typeable a) => Path as1 pathState -> Path (a ': as1) pathState

Wildcard :: forall (as1 :: [Type]). Path as1 'Open -> Path (Text ': as1) 'Closed

toInternalPath :: forall (as :: [Type]) (pathState :: PathState). Path as pathState -> PathInternal as

type Var a = Path '[a] 'Open

data AltVar a b

A variant of Either with a FromHttpApiData definition that tries both branches without a prefix. Useful to define routes with vars that should work with different types.

Constructors

AvLeft a

AvRight b

var :: (Typeable a, FromHttpApiData a) => Path '[a] 'Open

A route parameter

static :: String -> Path ('[] :: [Type]) 'Open

A static route piece

root :: Path ('[] :: [Type]) 'Open

The root of a path piece. Use to define a handler for "/"

wildcard :: Path '[Text] 'Closed

Matches the rest of the route. Should be the last part of the path.

(</>) :: forall (as :: [Type]) (bs :: [Type]) (ps2 :: PathState). Path as 'Open -> Path bs ps2 -> Path (Append as bs) ps2

pathToRep :: forall (as :: [Type]) (ps :: PathState). Path as ps -> Rep as

renderRoute :: forall (as :: [Type]). AllHave ToHttpApiData as => Path as 'Open -> HVect as -> Text

renderRoute' :: forall (as :: [Type]). AllHave ToHttpApiData as => Path as 'Open -> HVect as -> [Text]