Web.Spock.Core

Lauching Spock

runSpock :: Port -> IO Middleware -> IO ()

Run a Spock application using run from Network.Wai.Handler.Warp.

runSpockNoBanner :: Port -> IO Middleware -> IO ()

Like runSpock, but does not display the banner "Spock is running on port XXX" on stdout.

spockAsApp :: IO Middleware -> IO Application

Convert a middleware to an application. All failing requests will result in a 404 page

Spock's route definition monad

spockT :: MonadIO m => (forall a. m a -> IO a) -> SpockT m () -> IO Middleware

Create a raw spock application with custom underlying monad Use runSpock to run the app or spockAsApp to create a Wai.Application The first argument is request size limit in bytes. Set to Nothing to disable.

spockConfigT :: MonadIO m => SpockConfig -> (forall a. m a -> IO a) -> SpockT m () -> IO Middleware

Like spockT, but with additional configuration for request size and error handlers passed as first parameter.

type SpockT = SpockCtxT ()

Register routes with empty context. m is the base monad used by both registration and request actions; the final result is normally (). Pass a runner for m to spockT to build the middleware.

data SpockCtxT ctx (m :: Type -> Type) a

Register routes whose handlers use ActionCtxT ctx m. Registration runs when the application is built, while the registered handlers and prehooks run for matching requests. ctx is supplied by prehook, m provides base effects, and a is the registration result. lift enters m at setup time.

Defining routes

data Path (as :: [Type]) (pathState :: PathState)

root :: Path ('[] :: [Type]) 'Open

The root of a path piece. Use to define a handler for "/"

type Var a = Path '[a] 'Open

data AltVar a b

A variant of Either with a FromHttpApiData definition that tries both branches without a prefix. Useful to define routes with vars that should work with different types.

Constructors

AvLeft a

AvRight b

var :: (Typeable a, FromHttpApiData a) => Path '[a] 'Open

A route parameter

static :: String -> Path ('[] :: [Type]) 'Open

A static route piece

(<//>) :: forall (as :: [Type]) (bs :: [Type]) (ps :: PathState). Path as 'Open -> Path bs ps -> Path (Append as bs) ps

Combine two path components

wildcard :: Path '[Text] 'Closed

Matches the rest of the route. Should be the last part of the path.

Rendering routes

renderRoute :: forall (as :: [Type]). AllHave ToHttpApiData as => Path as 'Open -> HVectElim as Text

Render a route applying path pieces

Hooking routes

prehook :: forall t (m :: Type -> Type) ctx ctx'. (RouteM t, MonadIO m) => ActionCtxT ctx m ctx' -> t ctx' m () -> t ctx m ()

Specify an action that will be run before all subroutes. It can modify the requests current context

get :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb GET and the given route match

post :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb POST and the given route match

getpost :: forall (xs :: [Type]) t (m :: Type -> Type) ctx (ps :: PathState). (HasRep xs, RouteM t, Monad m, Monad (t ctx m)) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb GET/POST and the given route match

head :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb HEAD and the given route match

put :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb PUT and the given route match

delete :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb DELETE and the given route match

patch :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when the HTTP verb PATCH and the given route match

hookRoute :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => StdMethod -> Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when a standard HTTP verb and the given route match

hookRouteCustom :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Text -> Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when a custom HTTP verb and the given route match

hookAny :: forall t (m :: Type -> Type) ctx. (RouteM t, Monad m) => StdMethod -> ([Text] -> ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when a standard HTTP verb matches but no defined route matches. The full path is passed as an argument

hookAnyCustom :: forall t (m :: Type -> Type) ctx. (RouteM t, Monad m) => Text -> ([Text] -> ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when a custom HTTP verb matches but no defined route matches. The full path is passed as an argument

hookRouteAll :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run regardless of the HTTP verb

hookAnyAll :: forall t (m :: Type -> Type) ctx. (RouteM t, Monad m) => ([Text] -> ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run regardless of the HTTP verb and no defined route matches. The full path is passed as an argument

data StdMethod

HTTP standard method (as defined by RFC 2616, and PATCH which is defined by RFC 5789).

Since: http-types-0.2.0

Constructors

GET

POST

HEAD

PUT

DELETE

TRACE

CONNECT

OPTIONS

PATCH
Since: http-types-0.8.0

Adding Wai.Middleware

middleware :: forall t (m :: Type -> Type) ctx. (RouteM t, Monad m) => Middleware -> t ctx m ()

Hook wai middleware into Spock

Actions

module Web.Spock.Action

Config

data SpockConfig

Constructors

SpockConfig

Fields

sc_maxRequestSize :: Maybe Word64

Maximum request body size in bytes, checked as an action reads the body (including JSON, form parameters, and uploads). An unused body is not read or rejected based on its declared length. Exceeding the limit invokes sc_errorHandler with status 413.

sc_errorHandler :: Status -> ActionCtxT () IO ()

Error handler. Given status is set in response by default, but you can always override it with setStatus

sc_logError :: Text -> IO ()

Function that should be called to log errors.

sc_logging :: Maybe LoggingConfig

Optional request IDs and structured handleraccesserror events.

defaultSpockConfig :: SpockConfig

Default Spock configuration. No restriction on maximum request size; error handler simply prints status message as plain text and all errors are logged to stderr.

Internals

hookRoute' :: forall (xs :: [Type]) t (m :: Type -> Type) (ps :: PathState) ctx. (HasRep xs, RouteM t, Monad m) => SpockMethod -> Path xs ps -> HVectElim xs (ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when a HTTP verb and the given route match

hookAny' :: forall t (m :: Type -> Type) ctx. (RouteM t, Monad m) => SpockMethod -> ([Text] -> ActionCtxT ctx m ()) -> t ctx m ()

Specify an action that will be run when a HTTP verb matches but no defined route matches. The full path is passed as an argument

data SpockMethod

The SpockMethod allows safe use of http verbs via the MethodStandard constructor and StdMethod, and custom verbs via the MethodCustom constructor.

Constructors

MethodStandard !HttpMethod
Standard HTTP Verbs from StdMethod

MethodCustom !Text
Custom HTTP verbs represented as Text.

MethodAny
Match any HTTP verb

newtype HttpMethod

Constructors

HttpMethod

Fields

unHttpMethod :: StdMethod